Moved cluster code to cluster.js in helpers
