+++ backend/index.js
@@ -1,32 +1,10 @@
 require('dotenv').config();
-const cluster = require('cluster');
-const os = require('os');
-
-if (process.env.CLUSTER === 'yes') {
-  const numCPUs = os.cpus().length;
-
-  if (cluster.isMaster) {
-    console.log(`Starting app in cluster mode with ${numCPUs} workers`);
-    console.log(`Master ${process.pid} is running`);
-    for (let i = 0; i < numCPUs; i += 1) {
-      cluster.fork();
-    }
-
-    cluster.on('exit', (worker, code, signal) => {
-      console.log(`worker ${worker.process.pid} died with code: ${code} and signal: ${signal}`);
-      console.log('starting new worker');
-      cluster.fork();
-    });
-
-    return;
-  }
-  // else it's a worker. continue.
-  console.log(`Worker ${process.pid} started`);
-} else {
-  console.log('Starting app in non-cluster mode. (To start in cluster mode, pass CLUSTER=yes in config file)');
-}
-
 const http = require('http');
+
+const cluster = require('./helpers/cluster');
+
+if (cluster().isMaster) return;
+
 const app = require('./app');
 /**
  * Normalize a port into a number, string, or false.
